pyproject.toml
[project]
name = "ai-register"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "curl-cffi>=0.13.0",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tempmail-lol>=3.0.0",
]
